Refactor & Add Wallet link in User Profile
