+++ src/wallet/UserWalletSummary.less
@@ -35,5 +35,6 @@
     font-size: 18px;
     font-weight: bold;
     margin-left: auto;
+    text-align: right;
   }
 }
