+++ src/wallet/UserWalletTransactions.less
@@ -21,6 +21,10 @@
     margin-top: 5px;
   }
 
+  &__icon {
+    font-size: 35px !important;
+  }
+
   &__content {
     display: flex;
     flex-direction: column;
